requirements.txt
streamlit
pyyaml
faker